+++ src/Ext/AnimType/Body.cpp
@@ -110,6 +110,7 @@ void AnimTypeExt::ExtData::LoadFromINIFile(CCINIClass* pINI)
 	this->SplashAnims_PickRandom.Read(exINI, pID, "SplashAnims.PickRandom");
 	this->AttachedSystem.Read(exINI, pID, "AttachedSystem", true);
 	this->AltPalette_ApplyLighting.Read(exINI, pID, "AltPalette.ApplyLighting");
+	this->MakeInfantryOwner.Read(exINI, pID, "MakeInfantryOwner");
 }
 
 template <typename T>
@@ -143,6 +144,7 @@ void AnimTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->SplashAnims_PickRandom)
 		.Process(this->AttachedSystem)
 		.Process(this->AltPalette_ApplyLighting)
+		.Process(this->MakeInfantryOwner)
 		;
 }