+++ src/Ext/AnimType/Body.h
@@ -44,6 +44,7 @@ public:
 		Valueable<bool> SplashAnims_PickRandom;
 		Valueable<ParticleSystemTypeClass*> AttachedSystem;
 		Valueable<bool> AltPalette_ApplyLighting;
+		Valueable<OwnerHouseKind> MakeInfantryOwner;
 
 		ExtData(AnimTypeClass* OwnerObject) : Extension<AnimTypeClass>(OwnerObject)
 			, Palette { CustomPalette::PaletteMode::Temperate }
@@ -72,6 +73,7 @@ public:
 			, SplashAnims_PickRandom { false }
 			, AttachedSystem {}
 			, AltPalette_ApplyLighting { false }
+			, MakeInfantryOwner { OwnerHouseKind::Victim }
 		{ }
 
 		virtual ~ExtData() = default;